Extract chat workspace composer
